Add a script to invoke pgindent.
pgindent is the PostgreSQL equivalent of clang-format.
Unfortunately, pgindent is not aware of semantics.
pgindent takes a list of typedefs as input and uses the
list of typedefs to reformat the code. Therefore, if
you add a new typedef, you must add it to typedefs-noisepage.list.

Additionally, pgindent is not run on every commit, but is instead
run periodically before major PostgreSQL releases. Therefore, a
list of missing typedefs as of 13.4 was manually created. Then
this list is cat'd together with our typedefs, the old typedefs,
to produce the final list of typedefs necessary.